# Labels: Add DropMetricName function, used in PromQL