This function is called very frequently when executing PromQL functions, and we can do it much more efficiently inside Labels.

In the common case that `__name__` comes first in the labels, we simply re-point to start at the next label, which is nearly free.

`DropMetricName` is now so cheap I removed the cache - benchmarks show everything still goes faster.

<details><summary>Benchmarks</summary>
<p>

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
                                                                              │  before.txt  │             after.txt              │
                                                                              │    sec/op    │   sec/op     vs base               │
RangeQuery/expr=a_hundred,steps=100-8                                            487.7µ ± 0%   487.2µ ± 1%        ~ (p=1.000 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                  1.028m ± 1%   1.027m ± 2%        ~ (p=0.937 n=6)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                562.0m ± 1%   548.8m ± 3%   -2.35% (p=0.041 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-8                               97.80m ± 1%   97.49m ± 1%        ~ (p=0.699 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                  75.48m ± 1%   75.43m ± 2%        ~ (p=0.937 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-8                      44.87m ± 0%   45.07m ± 2%        ~ (p=0.485 n=6)
RangeQuery/expr=-a_hundred,steps=100-8                                           502.1µ ± 0%   495.1µ ± 0%   -1.39% (p=0.002 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                                3.045m ± 2%   3.062m ± 1%        ~ (p=0.394 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8               1.588m ± 1%   1.593m ± 0%        ~ (p=0.240 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8                1.923m ± 1%   1.935m ± 0%        ~ (p=0.065 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8            1.593m ± 3%   1.598m ± 1%        ~ (p=0.394 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=100-8                548.9µ ± 3%   552.6µ ± 1%        ~ (p=0.132 n=6)
RangeQuery/expr=abs(a_hundred),steps=100-8                                      1139.3µ ± 0%   949.8µ ± 1%  -16.63% (p=0.002 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-8    1.170m ± 0%   1.174m ± 0%        ~ (p=0.589 n=6)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-8           1.207m ± 1%   1.209m ± 1%        ~ (p=0.485 n=6)
RangeQuery/expr=sum(a_hundred),steps=100-8                                       612.6µ ± 0%   615.9µ ± 1%   +0.54% (p=0.002 n=6)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-8                           7.065m ± 6%   7.073m ± 1%        ~ (p=0.937 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-8                          9.132m ± 2%   9.150m ± 1%        ~ (p=0.937 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-8                                9.268m ± 5%   9.273m ± 0%        ~ (p=0.699 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=100-8                               7.036m ± 1%   7.086m ± 1%   +0.70% (p=0.041 n=6)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-8                     131.7m ± 0%   131.5m ± 1%        ~ (p=0.589 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=100-8                                   630.2µ ± 1%   634.4µ ± 0%   +0.66% (p=0.041 n=6)
RangeQuery/expr=topk(5,_a_hundred),steps=100-8                                   848.5µ ± 1%   844.9µ ± 1%        ~ (p=0.589 n=6)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=100-8            4.031m ± 1%   4.017m ± 0%   -0.34% (p=0.015 n=6)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-8                 1.161m ± 0%   1.148m ± 1%   -1.09% (p=0.002 n=6)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=100-8         21.30m ± 1%   21.11m ± 2%        ~ (p=0.093 n=6)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=100-8                  972.8µ ± 4%   962.8µ ± 0%        ~ (p=0.180 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=100-8                                1169.4µ ± 1%   960.6µ ± 1%  -17.86% (p=0.002 n=6)
geomean                                                                          3.759m        3.705m        -1.43%

                                                                              │  before.txt   │              after.txt              │
                                                                              │     B/op      │     B/op      vs base               │
RangeQuery/expr=a_hundred,steps=100-8                                            75.04Ki ± 0%   75.04Ki ± 0%        ~ (p=0.701 n=6)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                  126.1Ki ± 0%   100.3Ki ± 0%  -20.46% (p=0.002 n=6)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                3.514Mi ± 0%   3.487Mi ± 0%   -0.77% (p=0.002 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-8                               3.069Mi ± 1%   3.024Mi ± 1%        ~ (p=0.065 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                  3.050Mi ± 1%   3.041Mi ± 1%        ~ (p=0.699 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-8                      3.215Mi ± 1%   3.189Mi ± 1%   -0.79% (p=0.002 n=6)
RangeQuery/expr=-a_hundred,steps=100-8                                          104.43Ki ± 0%   78.62Ki ± 0%  -24.71% (p=0.002 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                                288.3Ki ± 0%   288.3Ki ± 0%        ~ (p=0.485 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8               680.3Ki ± 0%   680.2Ki ± 0%        ~ (p=0.394 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8                1.195Mi ± 0%   1.195Mi ± 0%        ~ (p=0.310 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8            680.3Ki ± 0%   680.3Ki ± 0%        ~ (p=0.853 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=100-8                127.6Ki ± 0%   127.6Ki ± 0%        ~ (p=0.851 n=6)
RangeQuery/expr=abs(a_hundred),steps=100-8                                       164.7Ki ± 0%   123.2Ki ± 0%  -25.20% (p=0.002 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-8    185.6Ki ± 0%   185.6Ki ± 0%        ~ (p=0.394 n=6)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-8           186.3Ki ± 0%   186.3Ki ± 0%        ~ (p=0.937 n=6)
RangeQuery/expr=sum(a_hundred),steps=100-8                                       136.8Ki ± 0%   136.7Ki ± 0%        ~ (p=0.814 n=6)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-8                           1.368Mi ± 0%   1.368Mi ± 0%        ~ (p=0.818 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-8                          3.772Mi ± 0%   3.772Mi ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-8                                3.772Mi ± 0%   3.772Mi ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=100-8                               1.368Mi ± 0%   1.368Mi ± 0%        ~ (p=0.937 n=6)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-8                     97.78Mi ± 0%   97.78Mi ± 0%        ~ (p=0.937 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=100-8                                   142.5Ki ± 0%   142.5Ki ± 0%        ~ (p=0.165 n=6)
RangeQuery/expr=topk(5,_a_hundred),steps=100-8                                   181.4Ki ± 0%   181.4Ki ± 0%        ~ (p=0.418 n=6)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=100-8            385.1Ki ± 0%   333.5Ki ± 0%  -13.42% (p=0.002 n=6)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-8                 188.5Ki ± 0%   162.7Ki ± 0%  -13.69% (p=0.002 n=6)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=100-8         1.739Mi ± 0%   1.452Mi ± 0%  -16.47% (p=0.002 n=6)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=100-8                  170.0Ki ± 0%   170.0Ki ± 0%        ~ (p=0.937 n=6)
RangeQuery/expr=timestamp(a_hundred),steps=100-8                                 625.4Ki ± 0%   583.8Ki ± 0%   -6.64% (p=0.002 n=6)
geomean                                                                          625.2Ki        595.2Ki        -4.80%

                                                                              │ before.txt  │              after.txt               │
                                                                              │  allocs/op  │  allocs/op   vs base                 │
RangeQuery/expr=a_hundred,steps=100-8                                           1.199k ± 0%   1.199k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                 1.944k ± 0%   1.644k ± 0%  -15.43% (p=0.002 n=6)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8               37.39k ± 0%   37.08k ± 0%   -0.81% (p=0.002 n=6)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-8                              37.29k ± 0%   36.99k ± 0%   -0.81% (p=0.002 n=6)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                 37.69k ± 0%   37.39k ± 0%   -0.79% (p=0.002 n=6)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-8                     37.38k ± 0%   37.08k ± 0%   -0.80% (p=0.002 n=6)
RangeQuery/expr=-a_hundred,steps=100-8                                          1.517k ± 0%   1.217k ± 0%  -19.78% (p=0.002 n=6)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                               2.997k ± 0%   2.997k ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8              2.989k ± 0%   2.989k ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8               3.149k ± 0%   3.149k ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8           2.989k ± 0%   2.989k ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=100-8               1.568k ± 0%   1.568k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=abs(a_hundred),steps=100-8                                      1.644k ± 0%   1.334k ± 0%  -18.86% (p=0.002 n=6)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-8   2.462k ± 0%   2.462k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-8          2.429k ± 0%   2.429k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=sum(a_hundred),steps=100-8                                      1.640k ± 0%   1.640k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-8                          15.15k ± 0%   15.15k ± 0%        ~ (p=1.000 n=6)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-8                         34.17k ± 0%   34.17k ± 0%        ~ (p=0.446 n=6)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-8                               34.17k ± 0%   34.17k ± 0%        ~ (p=0.994 n=6)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=100-8                              15.15k ± 0%   15.15k ± 0%        ~ (p=0.455 n=6)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-8                    574.3k ± 0%   574.3k ± 0%        ~ (p=0.502 n=6)
RangeQuery/expr=topk(1,_a_hundred),steps=100-8                                  1.763k ± 0%   1.763k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=topk(5,_a_hundred),steps=100-8                                  2.369k ± 0%   2.369k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=100-8           4.481k ± 0%   3.881k ± 0%  -13.40% (p=0.002 n=6)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-8                2.488k ± 0%   2.188k ± 0%  -12.06% (p=0.002 n=6)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=100-8        21.11k ± 0%   17.80k ± 0%  -15.64% (p=0.002 n=6)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=100-8                 1.601k ± 0%   1.601k ± 0%        ~ (p=1.000 n=6) ¹
RangeQuery/expr=timestamp(a_hundred),steps=100-8                                2.245k ± 0%   1.935k ± 0%  -13.81% (p=0.002 n=6)
geomean                                                                         6.058k        5.800k        -4.26%
¹ all samples are equal
```

</p>
</details> 

Originally written as part of #12304.